FACE_RECOGNIZE/requirements.txt
accelerate==0.21.0
addict==2.4.0
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
blendmodes==2022
celery==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
clean-fid==0.1.35
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clip==1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
deprecation==2.1.0
depth_anything @ https://github.com/huchenlei/Depth-Anything/releases/download/v1.0.0/depth_anything-2024.1.22.0-py2.py3-none-any.whl
depth_anything_v2 @ https://github.com/MackinationsAi/UDAV2-ControlNet/releases/download/v1.0.0/depth_anything_v2-2024.7.1.0-py2.py3-none-any.whl
diskcache==5.6.3
dlib==19.24.9
dsine @ https://github.com/sdbds/DSINE/releases/download/1.0.2/dsine-2024.3.23-py3-none-any.whl
easydict==1.13
einops==0.4.1
exceptiongroup==1.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
geffnet==1.0.2
gitdb==4.0.12
GitPython==3.1.32
glob2==0.5
gradio==3.41.2
gradio_client==0.5.0
h11==0.12.0
httpcore==0.15.0
httptools==0.6.4
httpx==0.24.1
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
inflection==0.5.1
iopath==0.1.9
Jinja2==3.1.6
joblib==1.5.1
jsonmerge==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.3
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
narwhals==1.40.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.2
omegaconf==2.2.3
onnxruntime-gpu==1.22.0
open-clip-torch==2.20.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
platformdirs==4.3.8
portalocker==3.1.1
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.7
psutil==5.9.5
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==1.9.4
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
resize-right==0.0.2
rpds-py==0.25.0
rtree==1.4.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.1.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spandrel==0.3.4
spandrel_extra_arches==0.1.1
starlette==0.26.1
stringzilla==3.12.5
svg.path==6.3
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.9.5
tinycss2==1.4.0
tokenizers==0.13.3
tomesd==0.1.3
tomli==2.2.1
torch==2.1.2+cu121
torchdiffeq==0.2.3
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.16.2+cu121
tqdm==4.67.1
trampoline==0.1.2
transformers==4.30.2
trimesh==4.6.10
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vhacdx==0.0.8.post2
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
xformers==0.0.23.post1
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.0
zipp==3.21.0
